+++ src/stock-time-series/dto/monthly-adjusted.dto.ts
@@ -1,0 +1,6 @@
+import { DataType } from '../../enum/datatype.enum';
+
+export type MonthlyAdjustedDTO = {
+  symbol: string;
+  datatype?: DataType;
+};