+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Daily adjusted time series integration
 - Weekly adjusted time series integration
+- Monthly adjusted time series integration
 
 ## [1.1.1]